Create collab note book
As a beginner in python, data science, ML, and DL.
This is my first experience in analyzing data.
It starts with a bitcoin dataset to find gross profit and visualize it with a simple plot graph of gross profit versus date. where gross profit is calculated using  [gross profit or loss = opening value - closing value of that day ].
The next data set is wine quality dataset in which based upon the count of the column values of the quality of wine computes the bar plot with the various quality labels on x-axis and their frequency values on y-axis.